Socket bonus and world php zones added
Socketed item tooltips now show the socket bonus and equipped items
also show if the socket bonus is active.
Realm tooltips now display the current status of the world pvp zones
(Wintergrasp and Tol Barad) and now sport an extended mode that
switches the world php zone display to the date time of the next battle
for the zone.